requirements.txt
tensorflow-gpu
keras
python-chess
numpy
tqdm
keras-tqdm